Add '.display' div to the HTML where to display the books.
Add 'Book' constructor to store all books in
'myLibrary' array.

Add 'addBookToLibrary' that takes arguments to create
books and store them in the 'myLibrary' array.

Add 'displayBooks' function that displays that loops
through the array and displays each book on the page,
by creating '.card' element and displaying the book
properties.

Style '.display' and '.card' in 'style.css' to be
visually pleasing.